pyproject.toml
[tool.poetry]
name = "ohota_na_skidki_bot"
version = "0.1.0"
description = "Telegram bot that tracks prices for you."
authors = ["Rasul Kireev <me@rasulkireev.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django-allauth = "^0.50.0"
python-webpack-boilerplate = "^1.0.0"
django-widget-tweaks = "^1.4.12"
mkdocs = "^1.4.2"
mkdocstrings = "^0.20.0"
mkdocs-material = "^9.0.12"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
django-structlog = "^8.1.0"
structlog = "^24.2.0"
django = "^5.0.6"
python-telegram-bot = "^21.3"
httpx = "^0.27.0"
beautifulsoup4 = "^4.12.3"
scrapingbee = "^2.0.1"

[tool.poetry.dev-dependencies]


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.1"
pylint-django = "^2.5.3"
pre-commit = "^3.2.1"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.23"

[tool.djlint]
profile="django"
ignore = "H031"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
